requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Data Validation & Settings
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality
black==23.12.0
pylint==3.0.3
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
